+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support cases where `type` is an array in JSONoid object conversion
 
+### Changed
+- Make JSONoid conversion helper methods private
+
 ## [0.5.3]
 ### Changed
 - Throw a more readable exception if `$ref` or `allOf` are found during conversion